Split up source_info.rs
